Updated doc to reflect current proposal...
Not as much a proposal at this point actually; this is the way I'm now
implementing it.  Although we're still in the 'RFC'/'look for horrible
lossage' stage, this is close to being set in stone unless we find
something horribly wrong with it.

Doc is still very light on detailed rationale and examples; I'd like
to subcontract that part of the writing and get on with code.



git-svn-id: http://svn.xiph.org/trunk/ogg@6719 0101bb08-14d6-0310-b084-bc0e0c8e3800