Update configure script to use new 'legal' names for cache vals (I hate you so much, autofoo)
Update ogg to check the return of all allocs for those on embedded
platforms with guarantees non-ovecommit.  Be aware that these checks
are useless on any modern desktop OS, but that embedded folks with no
MMU and a hard heap boundary will benefit.

Add one new call to bitpacker that will query a write-mode
oggpack_buffer to see if it ran out of memory at some point; there was
no other way to add the error reporting in an ABI-compatible manner
(and it also allows the writepacker to avoid extra checks; it will
just keep going after running out of memory without faulting and state
can be queried later).



git-svn-id: http://svn.xiph.org/trunk/ogg@16016 0101bb08-14d6-0310-b084-bc0e0c8e3800